+++ src/client/distributed_delete.cpp
