+++ src/client/entry_lookup_info_hash_utils.cpp
