+++ src/client/motherduck_catalog.cpp
