+++ src/client/motherduck_index_catalog_entry.cpp
