+++ src/client/motherduck_pragmas.cpp
