+++ src/client/motherduck_schema_catalog_entry.cpp
@@ -1,7 +1,6 @@
 #include "motherduck_schema_catalog_entry.hpp"
 
 #include "distributed_client.hpp"
-#include "distributed_protocol.hpp"
 #include "duckdb/catalog/catalog_entry/duck_index_entry.hpp"
 #include "duckdb/catalog/catalog_entry/duck_schema_entry.hpp"
 #include "duckdb/logging/logger.hpp"
