+++ src/client/motherduck_table_catalog_entry.cpp
