+++ src/include/client/distributed_client.hpp
@@ -3,7 +3,6 @@
 #pragma once
 
 #include "distributed_flight_client.hpp"
-#include "distributed_protocol.hpp"
 #include "duckdb.hpp"
 #include "duckdb/common/string.hpp"
 #include "duckdb/common/unique_ptr.hpp"
