+++ src/include/client/distributed_delete.hpp
