+++ src/include/client/motherduck_catalog.hpp
