+++ src/include/client/motherduck_index_catalog_entry.hpp
