+++ src/include/client/motherduck_pragmas.hpp
