+++ src/include/client/motherduck_schema_catalog_entry.hpp
