+++ src/include/client/motherduck_table_catalog_entry.hpp
