+++ src/server/distributed_flight_server.cpp
@@ -1,6 +1,5 @@
 #include "distributed_flight_server.hpp"
 
-#include "distributed_protocol.hpp"
 #include "duckdb/common/arrow/arrow_appender.hpp"
 #include "duckdb/common/arrow/arrow_converter.hpp"
 #include "duckdb/common/arrow/arrow_wrapper.hpp"
