+++ src/utils/time_utils.cpp
@@ -1,5 +1,5 @@
 #include "duckdb/common/types/timestamp.hpp"
-#include "time_utils.hpp"
+#include "utils/time_utils.hpp"
 
 namespace duckdb {
